Put serializers next to their views
