--- resources_portal/serializers/__init__.py
@@ -1,2 +1,0 @@
-from .material import MaterialSerializer
-from .user import CreateUserSerializer, UserSerializer
