+++ resources_portal/views/material.py
@@ -1,7 +1,21 @@
-from rest_framework import mixins, viewsets
+from rest_framework import mixins, serializers, viewsets
 
 from ..models import Material
-from ..serializers import MaterialSerializer
+
+
+class MaterialSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Material
+        fields = (
+            "id",
+            "url",
+            "pubmed_id",
+            "metadata",
+            "primary_contact",
+            "created_at",
+            "updated_at",
+        )
+        read_only_fields = ()
 
 
 class MaterialViewSet(mixins.ListModelMixin, viewsets.GenericViewSet):
