+++ resources_portal/views/user.py
@@ -1,9 +1,20 @@
-from rest_framework import mixins, viewsets
+from rest_framework import mixins, serializers, viewsets
 from rest_framework.permissions import AllowAny
 
 from ..models import User
 from ..permissions import IsUserOrReadOnly
-from ..serializers import CreateUserSerializer, UserSerializer
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = (
+            "id",
+            "username",
+            "first_name",
+            "last_name",
+        )
+        read_only_fields = ("username",)
 
 
 class UserViewSet(mixins.RetrieveModelMixin, mixins.UpdateModelMixin, viewsets.GenericViewSet):
@@ -14,6 +25,28 @@ class UserViewSet(mixins.RetrieveModelMixin, mixins.UpdateModelMixin, viewsets.G
     queryset = User.objects.all()
     serializer_class = UserSerializer
     permission_classes = (IsUserOrReadOnly,)
+
+
+class CreateUserSerializer(serializers.ModelSerializer):
+    def create(self, validated_data):
+        # call create_user on user object. Without this
+        # the password will be stored in plain text.
+        user = User.objects.create_user(**validated_data)
+        return user
+
+    class Meta:
+        model = User
+        fields = (
+            "id",
+            "username",
+            "password",
+            "first_name",
+            "last_name",
+            "email",
+            "auth_token",
+        )
+        read_only_fields = ("auth_token",)
+        extra_kwargs = {"password": {"write_only": True}}
 
 
 class UserCreateViewSet(mixins.CreateModelMixin, viewsets.GenericViewSet):
